// Frontend/src/models/PaymentMethod.ts
import BaseResponse from "./BaseResponse";

export interface PaymentMethodResponse extends BaseResponse {
    name: string;
    code: PaymentMethodType;
    status: number;
}

export interface PaymentMethodRequest {
    status: number;
}

export enum PaymentMethodType {
    CASH = "CASH",
    PAYPAL = "PAYPAL",
}